requirements.txt
dateutil
pillow
schedule
omxplayer-wrapper
python-vlc

